+++ src/main/java/github/renderbr/hytale/listeners/PlayerStateListener.java
@@ -4,8 +4,11 @@ import com.hypixel.hytale.event.EventRegistry;
 import com.hypixel.hytale.server.core.Message;
 import com.hypixel.hytale.server.core.event.events.player.PlayerDisconnectEvent;
 import com.hypixel.hytale.server.core.event.events.player.PlayerReadyEvent;
+import com.hypixel.hytale.server.core.event.events.player.PlayerSetupDisconnectEvent;
 import github.renderbr.hytale.AverageDiscord;
 import github.renderbr.hytale.config.obj.ChannelOutputTypes;
+import github.renderbr.hytale.config.obj.DiscordBridgeConfiguration;
+import github.renderbr.hytale.registries.ProviderRegistry;
 
 public class PlayerStateListener {
     public static void register(EventRegistry eventRegistry) {
@@ -14,12 +17,12 @@ public class PlayerStateListener {
     }
 
     public static void onPlayerJoin(PlayerReadyEvent event) {
-        AverageDiscord.instance.updateActivityPlayerCount();
+        AverageDiscord.instance.updateDiscordInformation();
         AverageDiscord.instance.SendMessageToType(ChannelOutputTypes.JOIN_LEAVE, Message.translation("server.bot.averagediscord.playerjoined").param("player", event.getPlayer().getDisplayName()).getAnsiMessage());
     }
 
     public static void onPlayerLeave(PlayerDisconnectEvent event) {
-        AverageDiscord.instance.updateActivityPlayerCount();
+        AverageDiscord.instance.updateDiscordInformation();
         AverageDiscord.instance.SendMessageToType(ChannelOutputTypes.JOIN_LEAVE, Message.translation("server.bot.averagediscord.playerleft").param("player", event.getPlayerRef().getUsername()).getAnsiMessage());
     }
 }